requirements.txt
asgiref==3.5.2
beautifulsoup4==4.10.0
certifi==2022.6.15
chardet==3.0.4
Django==3.2.9
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==2.10
numpy==1.22.3
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.24.0
sortedcontainers==2.2.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.47.0
urllib3==1.25.11
